Struct fedimint_wallet_server::Wallet

pub struct Wallet {
    pub(crate) cfg: WalletConfig,
    pub(crate) secp: Secp256k1<All>,
    pub(crate) btc_rpc: DynBitcoindRpc,
}

Fields
cfg: WalletConfigsecp: Secp256k1<All>btc_rpc: DynBitcoindRpc
Implementations

impl Wallet

pub async fn new(
cfg: WalletConfig,
db: Database,
task_group: &mut TaskGroup
) -> Result<Wallet>

pub async fn new_with_bitcoind(
cfg: WalletConfig,
db: Database,
bitcoind: DynBitcoindRpc,
task_group: &mut TaskGroup
) -> Result<Wallet, WalletError>

pub(crate) async fn save_peg_out_signatures<'a>(
&self,
dbtx: &mut ModuleDatabaseTransaction<'a>,
signatures: Vec<(PeerId, PegOutSignatureItem)>
)

pub(crate) fn sign_peg_out_psbt(
&self,
psbt: &mut PartiallySignedTransaction,
peer: &PeerId,
signature: &PegOutSignatureItem
) -> Result<(), ProcessPegOutSigError>

Try to attach signatures to a pending peg-out tx.

pub(crate) fn finalize_peg_out_psbt(
&self,
unsigned: UnsignedTransaction
) -> Result<PendingTransaction, ProcessPegOutSigError>

pub(crate) fn round_consensus(
last_height: u32,
items: Vec<(PeerId, RoundConsensusItem)>,
consensus_peers: &BTreeSet<PeerId>
) -> Result<RoundConsensus, Vec<PeerId>>

Calculates all the round items from peers, returning the same consensus for all peers or peers that should be banned for misbehaving

pub async fn current_round_consensus(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>
) -> Option<RoundConsensus>

pub async fn target_height(&self) -> u32

pub async fn consensus_height(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>
) -> Option<u32>

pub(crate) async fn sync_up_to_consensus_height<'a>(
&self,
dbtx: &mut ModuleDatabaseTransaction<'a>,
new_height: u32
)

pub(crate) async fn recognize_change_utxo<'a>(
&self,
dbtx: &mut ModuleDatabaseTransaction<'a>,
pending_tx: &PendingTransaction
)

Add a change UTXO to our spendable UTXO database after it was included in a block that we got consensus on.

pub(crate) async fn remove_rbf_transactions<'a>(
&self,
dbtx: &mut ModuleDatabaseTransaction<'a>,
pending_tx: &PendingTransaction
)

Removes the PendingTransaction and any transactions tied to it via RBF

pub(crate) async fn block_is_known(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>,
block_hash: BlockHash
) -> bool

pub(crate) async fn create_peg_out_tx(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>,
output: &WalletOutput
) -> Result<UnsignedTransaction, WalletError>

pub(crate) async fn available_utxos(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>
) -> Vec<(UTXOKey, SpendableUTXO)>

pub async fn get_wallet_value(
&self,
dbtx: &mut ModuleDatabaseTransaction<'_>
) -> Amount

pub(crate) fn offline_wallet(&self) -> StatelessWallet<'_>