Crate fedimint_wallet_server

Re-exports

pub use fedimint_wallet_common as common;

Modules

db

metrics 🔒

Structs

PendingTransaction

A peg-out tx that is ready to be broadcast with a tweak for the change UTXO

StatelessWallet 🔒

UnsignedTransaction

A PSBT that is awaiting enough signatures from the federation to becoming a PendingTransaction

Wallet

WalletInit

Functions

broadcast_pending_tx

calculate_pegin_metrics 🔒

calculate_pegout_metrics 🔒

nonce_from_idx

run_broadcast_pending_tx